save new changes

+++ delete-article.php
@@ -1,14 +1,19 @@
 <?php
-// $pdo = require_once './database.php';
-// $statementDeleteOne = $pdo->prepare('DELETE FROM article WHERE id = :id');
-$articleDB = require_once __DIR__ . '/database/model/articleDB.php';
 
-$_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-$id = $_GET['id'] ?? '';
-if ($id) {
-    // $statementDeleteOne->bindValue(':id', $id);
-    // $statementDeleteOne->execute();
-    $articleDB->deleteOne($id);
+require_once __DIR__ . '/database/database.php';
+$authDB = require_once __DIR__ . '/database/security.php';
+$currentUser = $authDB->isLoggedin();
+if ($currentUser) {
+    $articleDB = require_once __DIR__ . '/database/model/articleDB.php';
+    $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+    $id = $_GET['id'] ?? '';
+    if ($id) {
+        $article = $articleDB->fetchOne($id);
+        if ($article['author'] === $currentUser['id']) {
+            $articleDB->deleteOne($id);
+        }
+    }
+
 }
 header('Location: /');
